backend/requirements.txt
flask==3.1.2
flask-cors==6.0.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
numpy==1.26.4
protobuf==3.20.3
grpcio-status==1.46.3
pillow==12.0.0
google-genai==1.56.0
reportlab==4.4.7
python-dotenv==1.2.1
